+++ lca_algebraic/helpers.py
@@ -19,9 +19,15 @@ BIOSPHERE3_DB_NAME = 'biosphere3'
 _USER_DB = None
 
 def USER_DB() :
-    if _USER_DB is None :
-        raise Exception("User db name not set.  Please call SET_USER_DB('UserDbName') first ")
-    return _USER_DB
+    if _USER_DB is not None :
+        return _USER_DB
+    else:
+        # Try to guess it
+        candidates = list(key for key in bw.databases.keys() if key not in [ECOINVENT_DB_NAME(), BIOSPHERE3_DB_NAME])
+        if len(candidates) == 1 :
+            return candidates[0]
+
+        raise Exception("Unable to guess the name of the user DB. Please set it manually with SET_USEr_DB('YourDbName')")
 
 def SET_USER_DB(userDbName) :
     global _USER_DB